split summary into modules
