+++ table_z.py
@@ -1,0 +1,5 @@
+from table import out
+
+
+def build_table_z():
+    return out.copy()
